Revert "Adding date controllers to month and week"
