+++ README.md
@@ -1,26 +1,2 @@
 # my-groups-vac
 My Group's Vacation User Interface
-
-This project was built with Angular 8. 
-
-## Development
-
-To start development run the following command:
-
-```
-npm start
-```
-
-The above command servers the code on `http://localhost:4200` as well as keeps track of changes. When a change is saved the command will recompile the code and make it available to the browser. 
-
-
-## Production
-
-To run a production build run the following command:
-
-```
-npm run build:prod
-```
-
-The above command will use the angular.json file to compile and uglify the code into the `dist` directory. This command will not served the code. 
-
