+++ src/app/pages/login/login.component.spec.ts
@@ -1,7 +1,6 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LoginComponent } from './login.component';
-import { CustomMaterialModule } from '../../structure/material.modules';
 
 describe('LoginComponent', () => {
   let component: LoginComponent;
@@ -9,9 +8,7 @@ describe('LoginComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ LoginComponent ],
-      imports: [ CustomMaterialModule
-        ]
+      declarations: [ LoginComponent ]
     })
     .compileComponents();
   }));
@@ -26,4 +23,3 @@ describe('LoginComponent', () => {
     expect(component).toBeTruthy();
   });
 });
-*/
