+++ src/app/pages/signup/signup.component.spec.ts
@@ -1,7 +1,6 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SignupComponent } from './signup.component';
-import { CustomMaterialModule } from '../../structure/material.modules';
 
 describe('SignupComponent', () => {
   let component: SignupComponent;
@@ -9,9 +8,7 @@ describe('SignupComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ SignupComponent ],
-      imports: [ CustomMaterialModule
-        ]
+      declarations: [ SignupComponent ]
     })
     .compileComponents();
   }));
@@ -26,4 +23,3 @@ describe('SignupComponent', () => {
     expect(component).toBeTruthy();
   });
 });
-*/
