+++ src/app/pages/signup/signup.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material'
 
 @Component({
   selector: 'app-signup',
