--- src/app/pages/user/user.component.scss
@@ -1,12 +1,0 @@
-.user-vacation{
-  display: flex;
-}
-
-.user-group{
-    text-decoration: underline;    
-    color: #0000EE;
-}
-
-.user-group-selected{
-    background: #C5CAE0;
-}
