--- src/app/utils/constants.ts
@@ -1,18 +1,0 @@
-export const WEEK_DAYS:string[]=[
-  'Sunday',
-  'Monday', 'Tuesday','Wednesday',
-  'Thursday','Friday','Saturday'];
-
-export const MONTHS = [
-  'January','February','March',
-  'April','May','June','July',
-  'August','September','October',
-  'November','December'];
-
-
-export const MONTH_DAYS:number[]=[
-  1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
-  11, 12, 13, 14, 15, 16, 17, 18, 19, 20,
-  21, 22, 23, 24, 25, 26, 27, 28, 29, 30,
-  31
-];
