+++ src/app/utils/user/user.component.html
@@ -1,0 +1,1 @@
+<p>user works!</p>
