+++ src/app/utils/user/user.component.spec.ts
@@ -1,7 +1,6 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { UserComponent } from './user.component';
-import { CustomMaterialModule } from '../../structure/material.modules';
 
 describe('UserComponent', () => {
   let component: UserComponent;
@@ -9,9 +8,7 @@ describe('UserComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ UserComponent ],
-      imports: [ CustomMaterialModule
-        ]
+      declarations: [ UserComponent ]
     })
     .compileComponents();
   }));
@@ -26,4 +23,3 @@ describe('UserComponent', () => {
     expect(component).toBeTruthy();
   });
 });
-*/
