--- src/app/utils/vacation/day/day.component.html
@@ -1,7 +1,0 @@
-<div fxLayout="column" fxLayoutAlign="space-between stretch" class="base-container-brown day-container">
-  <div class="day-header" fxFlexAlign="center">{{header}}</div>
-  <div><hr></div>
-  <div *ngFor="let member of member_names" >
-    <div>{{member}}</div>
-  </div>
-</div>
