--- src/app/utils/vacation/day/day.component.scss
@@ -1,7 +1,0 @@
-.day-container{
-  padding: 8px;
-}
-.day-header{
-  font-weight: bold;
-  min-width: 150px;
-}
