--- src/app/utils/vacation/day/day.component.spec.ts
@@ -1,28 +1,0 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DayComponent } from './day.component';
-
-describe('DayComponent', () => {
-  let component: DayComponent;
-  let fixture: ComponentFixture<DayComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ DayComponent ],
-      imports: [ CustomMaterialModule
-        ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(DayComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-*/
