--- src/app/utils/vacation/month/month.component.html
@@ -1,24 +1,0 @@
-<div fxLayout="column" fxLayoutAlign="space-between start">
-    <div fxLayout="row" fxLayoutAlign="space-between center" class="month-controls-bar">
-        <mat-icon (click)="previousMonth()" class="month-icons">skip_previous</mat-icon> 
-        <span>Month: {{current_Month}}</span>
-        <mat-icon (click)="nextMonth()" class="month-icons">skip_next</mat-icon>
-    </div>
-
-    <div fxLayout="column" fxLayoutAlign="space-between stretch">
-        <div *ngFor="let week of six_week_days">
-            <div class="month-row">
-
-                <div fxLayout="row" fxLayoutAlign="space-between start">
-
-                    <div *ngFor="let day of week; index as d">
-                        <app-day [header]=" week_days[d+1] + ', ' + day.getDate() "
-                            [member_names]="getMembersOnVacation4Date(day)"></app-day>
-                    </div>
-
-                </div>
-            </div>
-        </div>
-    </div>
-
-</div>
