--- src/app/utils/vacation/month/month.component.spec.ts
@@ -1,28 +1,0 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { MonthComponent } from './month.component';
-
-describe('MonthComponent', () => {
-  let component: MonthComponent;
-  let fixture: ComponentFixture<MonthComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ MonthComponent ],
-      imports: [ CustomMaterialModule
-        ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(MonthComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-*/
