--- src/app/utils/vacation/vacation.component.html
@@ -1,11 +1,0 @@
-<div class="vacation-container" fxLayout="column" fxLayoutAlign="space-between center">
-  <mat-tab-group>
-    <mat-tab label="Week's view of the Group">
-      <app-week></app-week>
-    </mat-tab>
-    <mat-tab label="Month's vacation view">
-      <app-month></app-month>
-    </mat-tab>
-  </mat-tab-group>
-
-</div>
