--- src/app/utils/vacation/vacation.component.scss
@@ -1,3 +1,0 @@
-.vacation-container{
-  display: flex;
-}
