--- src/app/utils/vacation/vacation.component.spec.ts
@@ -1,28 +1,0 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { VacationComponent } from './vacation.component';
-
-describe('VacationComponent', () => {
-  let component: VacationComponent;
-  let fixture: ComponentFixture<VacationComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ VacationComponent ],
-      imports: [ CustomMaterialModule
-        ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(VacationComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-*/
