--- src/app/utils/vacation/vacation.component.ts
@@ -1,19 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-vacation',
-  templateUrl: './vacation.component.html',
-  styleUrls: ['./vacation.component.scss']
-})
-export class VacationComponent implements OnInit {
-
-  constructor() {
-
-  }
-
-  ngOnInit() {
-    
-  }
-
-
-}
