--- src/app/utils/vacation/week/week.component.html
@@ -1,33 +1,0 @@
-<div>
-    <div fxLayout="row" fxLayoutAlign="space-between center">
-        <div>
-            <mat-form-field>
-                <mat-label>Filter</mat-label>
-                <input matInput (keyup)="applyFilter($event)" placeholder="Example {{week_data[0]['Member']}}">
-            </mat-form-field>
-        </div>
-        <div fxLayout="row" fxLayoutAlign="space-between center">
-            <mat-icon (click)="previousWeek()" class="week-icons" >skip_previous</mat-icon>  <span>Week: {{current_week}}</span><mat-icon (click)="nextWeek()" class="week-icons" >skip_next</mat-icon> 
-        </div>
-    </div>
-    <table mat-table [dataSource]="dataSource" class="base-container-brown" matSort>
-        <ng-container [matColumnDef]="column" *ngFor="let column of headers">
-            <th mat-header-cell *matHeaderCellDef mat-sort-header> {{column}} </th>
-            <td mat-cell *matCellDef="let element">
-                <div *ngIf="element[column] == true" class="week-off">
-                    {{element[column]}}
-                </div>
-                <div *ngIf="element[column] == false" class="week-on">
-                    {{element[column]}}
-                </div>
-                <div *ngIf="element[column] != true && element[column] != false" class="week-first-column">
-                    {{element[column]}}
-                </div>
-
-            </td>
-        </ng-container>
-
-        <tr mat-header-row *matHeaderRowDef="headers"></tr>
-        <tr mat-row *matRowDef="let row; columns: headers;"></tr>
-    </table>
-</div>
