--- src/app/utils/vacation/week/week.component.spec.ts
@@ -1,28 +1,0 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { WeekComponent } from './week.component';
-
-describe('WeekComponent', () => {
-  let component: WeekComponent;
-  let fixture: ComponentFixture<WeekComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ WeekComponent ],
-      imports: [ CustomMaterialModule
-        ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(WeekComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-*/
